UC-12 refactor: rename Posts slice to Post
