+++ src/app/store/store.ts
@@ -3,18 +3,18 @@ import { configureStore, combineReducers } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/query/react'
 import { signInReducer } from '@/src/features/auth/signIn/model/signInSlice'
 import { loadState, saveState } from '@/src/shared/lib/localstorage'
-import { viewPostsApi } from '@/src/entities/Posts/service/api/viewPost.api'
+import { viewPostApi } from '@/src/entities/Post'
 
 const rootReducer = combineReducers({
   [baseApi.reducerPath]: baseApi.reducer,
-  [viewPostsApi.reducerPath]: viewPostsApi.reducer,
+  [viewPostApi.reducerPath]: viewPostApi.reducer,
 
   signIn: signInReducer,
 })
 export const store = configureStore({
   reducer: rootReducer,
   middleware: getDefaultMiddleware =>
-    getDefaultMiddleware().concat(baseApi.middleware).concat(viewPostsApi.middleware),
+    getDefaultMiddleware().concat(baseApi.middleware).concat(viewPostApi.middleware),
   preloadedState: loadState(),
 })
