+++ src/entities/Post/index.ts
@@ -1,0 +1,2 @@
+export { viewPostApi, useViewPostsQuery } from './api/viewPost.api'
+export { Post } from './ui/Post'
