+++ src/entities/Post/lib/truncateDescription.ts
