+++ src/entities/Post/model/types.ts
