+++ src/entities/Post/ui/AvatarImageLink.tsx
