+++ src/entities/Post/ui/post-lower/DescriptionPanel.tsx
