+++ src/entities/Post/ui/post-lower/LikesPanel.tsx
