+++ src/entities/Post/ui/post-lower/index.ts
