+++ src/entities/Post/ui/post-mediafiles/index.ts
