--- src/entities/Posts/index.ts
@@ -1,2 +1,0 @@
-export { viewPostsApi, useViewPostsQuery } from './api/viewPost.api'
-export { Post } from './ui/Post'
